added User and init spring security dependency maven

+++ Lab-1/src/main/java/evgesha/blps/lab1/repository/CategoryRepository.java
@@ -3,6 +3,6 @@ package evgesha.blps.lab1.repository;
 import evgesha.blps.lab1.entity.Category;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface CategoryRepository extends JpaRepository<Category, Integer> {
+public interface CategoryRepository extends JpaRepository<Category, Long> {
     public Category getCategoryByName(String name);
 }
