+++ Lab-1/src/main/java/evgesha/blps/lab1/repository/IngredientRepository.java
@@ -3,7 +3,7 @@ package evgesha.blps.lab1.repository;
 import evgesha.blps.lab1.entity.Ingredient;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface IngredientRepository extends JpaRepository<Ingredient, Integer> {
+public interface IngredientRepository extends JpaRepository<Ingredient, Long> {
