+++ Lab-1/src/main/java/evgesha/blps/lab1/repository/TargetRepository.java
@@ -3,7 +3,7 @@ package evgesha.blps.lab1.repository;
 import evgesha.blps.lab1.entity.Target;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface TargetRepository extends JpaRepository<Target, Integer> {
+public interface TargetRepository extends JpaRepository<Target, Long> {
 
     public Target getTargetByName(String name);
